#157969605 modify request and user endpoints
